### 2023-09-04 e0258eb
Fix affilation and reference
