+++ paper.bib
@@ -51,7 +51,8 @@
   booktitle={Proceedings of the 14th python in science conference},
   number={130-136},
   year={2015},
-  organization={Citeseer}
+  organization={Citeseer},
+  doi={10.25080/majora-7b98e3ed-013},
 }
 
 @Article{Hunter:2007,
@@ -78,7 +79,8 @@
   pages={40--51},
   issn={1521-9615},
   year={2011},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/mcse.2011.35}
 }
 
 @conference{Kluyver2016jupyter,
