### 2023-09-04 40ec218
Add index for affilation
