### 2023-09-11 #5802 @lys-devel
Hi @emdupre, thank you for the editorial effort.
I suggest potential reviewers who may be interested in the reviewing my software.

- [dxm447](https://reviewers.joss.theoj.org/reviewers/1488)
- [jonatanalvelid](https://github.com/jonatanalvelid)
- [kuadrat](https://github.com/kuadrat)
