### 2023-09-18 #5802 @kuadrat
Dear @emdupre , I would be happy to review this exciting submission!

However, I am the developer of [data-slicer](https://joss.theoj.org/papers/10.21105/joss.02969), a 3D data visualization and analysis tool. On quick inspection, `data-slicer` appears to have significant overlap with `lys`, both in functionality and targeted use-case.

I didn't find any mention of this being a problem in terms of conflict of interest or similar in the reviewing guideline. Nevertheless, I thought I'd disclose it here and ask you for confirmation that this does not disqualify me from reviewing `lys`.
