### 2023-09-19 #5802 @emdupre
Thank you for your enthusiastic response, @kuadrat !

We do not consider it as a conflict-of-interest to work on a related project, provided there is no co-development ; rather, this is likely to bring especially valuable perspective to your review.

That being said, if you feel that you cannot provide an impartial review _because_ of your work on `data-slicer` (e.g., because you feel that the two projects are in competition), then this would likely constitute a conflict-of-interest. Please let me know what you think, and thank you again !
