### 2023-09-20 #5802 @kuadrat
Thanks for your response, @emdupre .

I personally believe that I am able to review this project impartially.
1. The goal in developing `data-slicer` was to make the lives of fellow scientists easier. In that sense, if a tool comes along that better accomplishes what `data-slicer` could also be used for, that's completely in the spirit of `data-slicer`, even if it itself becomes redundant. As you stated, I'd be happy to bring in my experience and perspective to `lys`.
2. Realistically speaking, I don't have any resources to continue development and maintenance of `data-slicer`. Therefore, I would be happy to see an actively maintained alternative becoming available.

Despite this, I imagine that it could be perceived by others (maybe even the author, @lys-devel ) that the projects are competing and I would therefore make a bad choice of referee. Still, even in this case and given that papers in JOSS aren't typically rejected but instead the referee's are asked to make concrete suggestions for improvement, it seems to me that I should be eligible to review this. I would like to leave the final call to you, though.
