### 2023-09-22 #5802 @emdupre
👋  Hi @ThFriedrich @pr4deepr @ziatdinovmax,

Would you be willing to review _lys: interactive multi-dimensional data analysis and visualization platform_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/lys-devel/lys) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.05802/joss.05802/10.21105.joss.05802.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you for considering !
