### 2023-09-22 #5802 @ziatdinovmax
Will be happy to review it!

On Fri, Sep 22, 2023 at 12:44 PM Elizabeth DuPre ***@***.***>
wrote:

> 👋 Hi @ThFriedrich <https://github.com/ThFriedrich> @pr4deepr
> <https://github.com/pr4deepr> @ziatdinovmax
> <https://github.com/ziatdinovmax>,
>
> Would you be willing to review *lys: interactive multi-dimensional data
> analysis and visualization platform* for JOSS (the Journal of Open Source
> Software) <https://joss.theoj.org/>?
>
> You can learn more about reviewing for JOSS -- including our conflict of
> interest policy -- here
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>.
>
> The review takes place on GitHub and focuses on the software
> <https://github.com/lys-devel/lys> and this short paper:
>
> https://github.com/openjournals/joss-papers/blob/joss.05802/joss.05802/10.21105.joss.05802.pdf
>
> If you're available to review this work, please let me know and I'll add
> you as a reviewer. Once I have sufficient reviewers I'll open a dedicated
> review issue.
>
> Thank you for considering !
>
