### 2023-09-22 #5869 @editorialbot
[REVIEW]: lys: interactive multi-dimensional data analysis and visualization platform
**Submitting author:** @lys-devel (<a href="http://orcid.org/0000-0002-3010-9475">Asuka Nakamura</a>)
**Repository:** https://github.com/lys-devel/lys
**Branch with paper.md** (empty if default branch): review
**Version:** v0.3.3_zenodo
