### 2023-09-26 lys-devel/lys#3 @kuadrat
unittest throws "FileNotFoundError"s
Following [the instructions in the documentation](openjournals/joss-reviews/issues/5869), I carry out:
```
python -m unittest discover
```
in the `lys` code directory.

This results in an infinite loop of Traceback printed on console, at the core of which seems to be a `FileNotFoundError` to the file `lys/.lys/debug.log`.

Any idea what could be done to get the unittest to run as intended?

Part of openjournals/joss-reviews/issues/5869.
