### 2023-09-26 lys-devel/lys#5 @kuadrat
Tests for software integrity
From my perspective, `lys` scores *OK* in terms of tests for core functionality, as outlined in the [JOSS guidelines here](https://joss.readthedocs.io/en/latest/review_criteria.html#tests): The documentation contains step-by-step instructions for testing the core functionality in the form of a tutorial.

Instructions for a unittest for contributors are also given in the documentation (however, see #3) .

The following is therefore not a publication-critical issue, but rather a strongly encouraged suggestion to add automated tests by which users (not collaborators) can verify their installation. I would suggest the use of `pytest` and `pytest-qt`.

Part of https://github.com/openjournals/joss-reviews/issues/5869.
