### 2023-09-26 lys-devel/lys#6 @kuadrat
"State of the field" missing
According to the JOSS guidelines [a discussion of the state of the field is required](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper) in the submitted paper.

Comparing the pros and cons of `lys` to other existing software with similar functionality would not only check this box, but also help a reader of the documentation or the paper understand better what `lys` can and cannot do. As such, I believe that the [landing page of the `lys` documentation](https://lys-devel.github.io/lys/index.html) could also benefit from such a discussion.

Suggested related software solutions that came to mind when reading about and using `lys`:
- IGOR Pro: mature system for mixed GUI and CUI multidimensional data analysis. Not open-source.
- [data-slicer](github.com/kuadrat/data-slicer) (of which I am the author): Multidimensional data inspection and analysis framework with the GUI + CUI tool `pit`. Leans more on the CUI side and is more geared towards quick data inspection rather than data analysis and creation of quality figures.
- pyqtgraph: The underlying library for many of `lys`' and `data-slicer`s functionality. No GUI, requires at least intermediate python knowledge.

Part of https://github.com/openjournals/joss-reviews/issues/5869.
