### 2023-09-26 lys-devel/lys#7 @kuadrat
Typos in paper
A few typos I found in the compiled paper proof:

1. Line 6: the last word (*the*) is superfluous
2. Line 31: intera**c**tive
3. Line 50: *contain* (not *contains*)
4. Lines 87 and 93: functiona**l**ities

Part of openjournals/joss-reviews/issues/5869.
