### 2023-09-26 lys-devel/lys#10 @kuadrat
Documentation of proc.py / Understanding shell in lys
The overall documentation appears quite complete to my eye.
However, I would have liked to read more on the `proc.py` file which is used in the first tutorial. Unfortunately, the section [Understanding shell in lys](https://lys-devel.github.io/lys/tutorials/shell.html#understanding-shell-in-lys) appears to be under construction.
I think the package would benefit from completion of this section.

Part of https://github.com/openjournals/joss-reviews/issues/5869.
