### 2023-09-29 lys-devel/lys#11 @pr4deepr
system requirements on github repo
Could you state system requirements in the Github repo?
I am using a windows machine and although I have WSL, not everyone will.

As part of JOSS review: https://github.com/openjournals/joss-reviews/issues/5869
