### 2023-09-29 lys-devel/lys#12 @pr4deepr
installation issues (Windows 11; WSL2)
Hi
  I'm running Windows 11 and trying to use WSL2.

It appears there needs to be additional configuration that needs to be done.

Currently, when I follow the instructions on installation website and run lys, I get:

`/mnt/d/.venv/lys/bin/python3: No module named lys.__main__; 'lys' is a package and cannot be directly executed
`

I'll try conda instead of venv and get back to you

Cheers
Pradeep

https://github.com/openjournals/joss-reviews/issues/5869
