### 2023-09-29 1925cad
Update issue templates
Add template for JOSS review
