+++ .github/ISSUE_TEMPLATE/joss-review-issue.md
@@ -1,0 +1,10 @@
+---
+name: JOSS review issue
+about: 'Please use this template (label is set) for review issue of JOSS. '
+title: Title of the issue
+labels: JOSS_Review
+assignees: lys-devel
+
+---
+
+
