### 2023-09-29 lys-devel/lys#12 @pr4deepr
So, with WSL2, I had to configure a few things:

* Update WSL
* I ran the xeyes application to test the gui, but it wouldn't start
* It suggested:  `sudo apt install x11-apps`
* After installation, when running lys, I got this error
```
Traceback (most recent call last):
  File "<frozen runpy>", line 189, in _run_module_as_main
  File "<frozen runpy>", line 148, in _get_module_details
  File "<frozen runpy>", line 112, in _get_module_details
  File "/home/pradeepr/miniconda3/envs/lys_test/lib/python3.11/site-packages/lys/__init__.py", line 1, in <module>
    from . import Qt
  File "/home/pradeepr/miniconda3/envs/lys_test/lib/python3.11/site-packages/lys/Qt/__init__.py", line 11, in <module>
    from qtpy import QtWidgets, QtGui
  File "/home/pradeepr/miniconda3/envs/lys_test/lib/python3.11/site-packages/qtpy/QtWidgets.py", line 36, in <module>
    from PyQt5.QtWidgets import *
ImportError: libGL.so.1: cannot open shared object file: No such file or directory
```
* Installed libgl1: `sudo apt-get install libgl1`
Got this error when starting lys:

```
qt.qpa.plugin: Could not load the Qt platform plugin "xcb" in "" even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.

Available platform plugins are: eglfs, linuxfb, minimal, minimalegl, offscreen, vnc, wayland-egl, wayland, wayland-xcomposite-egl, wayland-xcomposite-glx, webgl, xcb.
```

* Ran these [two commands](https://askubuntu.com/a/1460390):
   * export QT_QPA_PLATFORM="xcb"
   * sudo apt install libxcb-xinerama0 libqt5x11extras5

When I start lys, I get a GUI and this output:
```

(lys_test) pradeepr@PRADEEP:/mnt/d/.venv$ python -m lys
Error in sys.excepthook:
object address  : 0x7fb0e970eaa0
object refcount : 1
object type     : 0x8718c0
object type name: SystemError
object repr     : SystemError("<method 'emit' of 'PyQt5.QtCore.pyqtBoundSignal' objects> returned a result with an exception set")
lost sys.stderr

Original exception was:
object address  : 0x7fb0e970ea40
object refcount : 3
object type     : 0x871720
object type name: TypeError
object repr     : TypeError("setFixedHeight(self, h: int): argument 1 has unexpected type 'float'")
lost sys.stderr
```

![image](https://github.com/lys-devel/lys/assets/13831458/f9da031c-a7fd-494d-92c4-37b7f74944a4)

Does this look right??
I haven't tested it yet.

CHeers
Pradeep
