### 2023-09-29 lys-devel/lys#12 @lys-devel
Thank you for the comment and trying to fix the problem.

The screenshot of GUI looks fine although I concern error message. The problem may be due to the Qt setting as you tried.
Could you let me know from which step you can launch `xeye` for the future troubleshooting?
lys will not work at least when xeye cannot be launched.

I will add troubleshooting page in documents and transfer it to this issue.
