### 2023-09-29 lys-devel/lys#11 @lys-devel
Thank you for the comment.
I added the statements regarding system requirements in the [top page](https://github.com/lys-devel/lys).
For maintainability, the detailed system requirements are described only in the [installation page](https://lys-devel.github.io/lys/install.html).
