### 2023-10-02 173f4f0
Add source installation page in docs #5
