+++ docs_src/contributing.rst
@@ -12,7 +12,7 @@ We really thank you for the input! We are welcome to most of feedbacks such as
 Report bugs
 ---------------------------
 
-We use GitHub issues to track public bugs. Report a bug by opening a new issue, which includes
+We use GitHub issues (https://github.com/lys-devel/lys/issues) to track public bugs. Report a bug by opening a new issue, which includes
 
 - A quick summary and/or background
 - Steps to reproduce
@@ -29,16 +29,14 @@ Pull requests are the best way to propose changes to the code. We actively welco
 1. Fork the repo and create your branch from `main`.
 2. Make sure your code follows pep8 (We recommend to use autopep8).
 3. If you've added code that should be tested, add tests.
-4. Ensure the unittest passes.
+4. Ensure the :doc:`tutorials/test` passes.
 5. If you've changed APIs, update the documentation. We use sphinx and reStructuredText for documentation.
 6. Issue that pull request in GitHub.
 
-Unittest
+Automated test
 ---------------
 
-Go to lys directory, and run the code below::
-
-    python -m unittest discover
+See :doc:`tutorials/test` for automated test.
 
 License
 ------------------------
