+++ docs_src/tutorials/tutorial.rst
@@ -30,4 +30,5 @@ Tips:
 .. toctree::
    :maxdepth: 1
 
+   test
    loadVector
