### 2023-10-02 lys-devel/lys#5 @lys-devel
Thank you @kuadrat for the comment for the automated tests.
I think it is not conventional to include automated tests in pypi wheel package as discussed [here](https://stackoverflow.com/questions/75561834/how-to-include-tests-from-root-project-directory-in-pip-package-such-that-other). Therefore I added the document to install lys from source and to execute automated test in [installation page](https://lys-devel.github.io/lys/install.html).
Do you think these modifications are enough for users to verify their installation?

In addition, I will use `pytest` for automated testing in future development, as you suggested. The [automated test](https://lys-devel.github.io/lys/tutorials/test.html) page in document is changed to use `pytest` and I have confirmed the present tests can be passed by using `pytest`.
