### 2023-10-02 1abbe22
Fix formula for residual sum in fitting #2
