+++ paper.md
@@ -20,17 +20,17 @@ bibliography: paper.bib
 
 # Summary
 
-`lys` is a Python-based multi-dimensional data analysis and visualization platform based on the several popular libraries such as `numpy`, `dask`, `matplotlib`, and `Qt`. It provides  graphical user interfaces (GUI) to intuitively and flexibly manipulate data array and graphics. In particular, multidimensional ($\ge$ 3D) data can be easily analyzed and visualized via optimized GUI interface, `MultiCut`. All the processes in `MultiCut` are automatically calculated in parallel using `dask` arrays, providing fast data analysis on high-performance computers (HPCs). These GUIs in `lys` offer interactive and low-code data analysis for users not familiar with the application programming interfaces (APIs) of the scientific Python ecosystems. In addition to the user-friendly GUIs, `lys` also provides flexibility for experts through its character user interface (CUI). Arbitrary Python code can be executed in the extended Python interpreter system, which can be integrated into `MultiCut`. The hybrid GUI/CUI architecture in `lys` enables an intuitive, low-code, parallel, flexible, and extensible multi-dimensional data analysis for a broad range of users. 
+`lys` is a Python-based multi-dimensional data analysis and visualization platform based on several popular libraries such as `numpy`, `dask`, `matplotlib`, and `Qt`. It provides  graphical user interfaces (GUI) to intuitively and flexibly manipulate data array and graphics. In particular, multidimensional ($\ge$ 3D) data can be easily analyzed and visualized via optimized GUI interface, `MultiCut`. All the processes in `MultiCut` are automatically calculated in parallel using `dask` arrays, providing fast data analysis on high-performance computers (HPCs). These GUIs in `lys` offer interactive and low-code data analysis for users not familiar with the application programming interfaces (APIs) of the scientific Python ecosystems. In addition to the user-friendly GUIs, `lys` also provides flexibility for experts through its character user interface (CUI). Arbitrary Python code can be executed in the extended Python interpreter system, which can be integrated into `MultiCut`. The hybrid GUI/CUI architecture in `lys` enables an intuitive, low-code, parallel, flexible, and extensible multi-dimensional data analysis for a broad range of users. 
 
 # Statement of need
 
-Data analysis and visualization are indispensable parts of scientific research. Understanding the experimental and simulated data deeply is essential for extracting complex phenomena behind them. To this end, intuitive and fast analysis/visualization is highly required. In the scientific Python ecosystem, flexible and fast analysis of numerical arrays has been done by several popular libraries such as `numpy`[@harris2020array], `scipy`[@2020SciPy-NMeth], `dask`[@rocklin2015dask] combined with visualization tools such as `matplotlib`[@Hunter:2007], `pyqtgraph`[@pyqtgraph] and `Mayavi`[@ramachandran2011mayavi]. Development of `Jupyter Notebook` [@Kluyver2016jupyter] and related libraries further enhance the capability of interactive data analysis. However, most of these libraries require users to be familiar with low-level application programming interfaces (APIs), which prevents intuitive analysis and visualization. In particular, when the analyzed data is a more than 3-dimensional array, even simple interative visualization requires Python code of tens of lines. Such on-demand analysis and visualization programs should be modified (and tested) when the data change, although a very similar process is frequently applied to data with different dimensions. Furthermore, the code for such an analysis should be preserved to guarantee the reproducibility of scientific results.
[…]
