+++ docs/_sources/install.rst.txt
@@ -33,7 +33,7 @@ Installation from pip (recommended)
 
     pip install lys-python
 
-3. Start lys by the commande below. Note that the current directory of the system is used as the working directory of lys::
+3. Start lys by the command below. Note that the current directory of the system is used as the working directory of lys::
 
     python -m lys
 
@@ -62,7 +62,7 @@ In particular, this is recommended if you want to use :doc:`tutorials/test`.
 
 4. (Optional) If you want to check installation, go to :doc:`tutorials/test`.
 
-5. Start lys by the commande below. Note that the current directory of the system is used as the working directory of lys::
+5. Start lys by the command below. Note that the current directory of the system is used as the working directory of lys::
 
     python -m lys
 
@@ -75,8 +75,7 @@ Library version
 
 We confirmed that *lys* works well under the environment below. If *lys* does not work, try the library versions below.
 
-- OS: AlmaLinux 8.3
-- Python 3.6.8
+Python 3.6.8
 
 - numpy 1.19.5
 - scipy 1.5.4
@@ -88,4 +87,18 @@ We confirmed that *lys* works well under the environment below. If *lys* does no
 - PyQt5 5.15.6
 - qtpy 2.0.1
 - autopep8 1.6.0
+
+
+Python 3.11.5
+
+- numpy 1.26.0
+- scipy 1.11.3
+- opencv-python-headless 4.8.1.78
+- dask 2023.9.3
+- dask-image 2023.8.1
+- matplotlib 3.8.0
+- pyqtgraph 0.13.3
+- PyQt5 5.15.9
+- qtpy 2.4.0
+- autopep8 2.0.4
 
