### 2023-10-02 lys-devel/lys#12 @lys-devel
Hi @pr4deepr ,
I have checked `lys` in Python 3.11 and found several errors because of change in dependent libraries.
At least, I removed one of the errors you got. Please let me know you have still some error.
To enable bug-fixed version, please install `lys` from source code by following [updated installation page](https://lys-devel.github.io/lys/install.html)
