### 2023-10-02 2adad81
Update document regarding shell in lys #10
