### 2023-10-02 lys-devel/lys#10 @lys-devel
Thank you for the important suggestion.
I have updated the document [Understanding shell in lys](https://lys-devel.github.io/lys/tutorials/shell.html), which describes how the Python interpreter in lys works.
