### 2023-10-03 59edd23
Add state of the field in the manuscript #6
