+++ paper.bib
@@ -217,4 +217,18 @@ pages  ="27-39",
 publisher  ="The Royal Society of Chemistry",
 doi  ="10.1039/D2FD00062H",
 url  ="http://dx.doi.org/10.1039/D2FD00062H",
-abstract  ="Ultrafast optical control of strain is crucial for the future development of nanometric acoustic devices. Although ultrafast electron microscopy has played an important role in the visualization of strain dynamics in the GHz frequency region{,} quantitative strain evaluation with nm × ps spatio-temporal resolution is still challenging. Five-dimensional scanning transmission electron microscopy (5D-STEM) is a powerful technique that measures time-dependent diffraction or deflection of the electron beam at the respective two-dimensional sample positions in real space. In this paper{,} we demonstrate that convergent beam electron diffraction (CBED) measurements using 5D-STEM are capable of quantitative time-dependent strain mapping in the nm × ps scale. We observe the generation and propagation of acoustic waves in a nanofabricated silicon thin plate of 100 nm thickness. The polarization and amplitude of the acoustic waves propagating in the silicon plate are quantitatively determined from the CBED analysis. Further Fourier-transformation analysis reveals the strain distribution in the momentum-frequency space{,} which gives the dispersion relation in arbitrary directions along the plate. Versatility of 5D-STEM-CBED analysis enables quantitative strain mapping even in complex nanofabricated samples{,} as demonstrated in this study."}
+abstract  ="Ultrafast optical control of strain is crucial for the future development of nanometric acoustic devices. Although ultrafast electron microscopy has played an important role in the visualization of strain dynamics in the GHz frequency region{,} quantitative strain evaluation with nm × ps spatio-temporal resolution is still challenging. Five-dimensional scanning transmission electron microscopy (5D-STEM) is a powerful technique that measures time-dependent diffraction or deflection of the electron beam at the respective two-dimensional sample positions in real space. In this paper{,} we demonstrate that convergent beam electron diffraction (CBED) measurements using 5D-STEM are capable of quantitative time-dependent strain mapping in the nm × ps scale. We observe the generation and propagation of acoustic waves in a nanofabricated silicon thin plate of 100 nm thickness. The polarization and amplitude of the acoustic waves propagating in the silicon plate are quantitatively determined from the CBED analysis. Further Fourier-transformation analysis reveals the strain distribution in the momentum-frequency space{,} which gives the dispersion relation in arbitrary directions along the plate. Versatility of 5D-STEM-CBED analysis enables quantitative strain mapping even in complex nanofabricated samples{,} as demonstrated in this study."
+}
+
+@article{Kramer2021, 
+doi = {10.21105/joss.02969}, 
+url = {https://doi.org/10.21105/joss.02969}, 
+year = {2021}, 
+publisher = {The Open Journal}, 
+volume = {6}, 
+number = {60}, 
+pages = {2969}, 
+author = {Kevin Kramer and Johan Chang}, 
+title = {Visualization of Multi-Dimensional Data -- The data-slicer Package}, 
+journal = {Journal of Open Source Software} 
+}
