### 2023-10-03 d2476b0
Modify docs to compare lys with other softwares #6
