### 2023-10-03 lys-devel/lys#6 @lys-devel
Thank you for the comment.
Following your suggestion, I added "state of the field" in the manuscript (last paragraph in `Overview` section). I also changed `Statement of need` (line 22-25, 29-31 in the new manuscript)

In addition, I also added the comparison between lys and other software in [landing page](https://lys-devel.github.io/lys/index.html). I appreciate your suggestion because it greatly benefits to readers and users of `lys`.
