### 2023-10-04 lys-devel/lys#3 @kuadrat
I have confirmed that `pytest` according to the new instructions [here](https://lys-devel.github.io/lys/tutorials/test.html) work as described (except for a couple of DeprecationWarnings regarding "invalid escape sequences" (\p, \e)).
