### 2023-10-06 lys-devel/lys#6 @lys-devel
Thank you for the fruitful comment.
- I mention advantages of open source software in the manuscript (line 99) and [document](https://lys-devel.github.io/lys/index.html).
- I added the screenshot in [README](https://github.com/lys-devel/lys) and [document](https://lys-devel.github.io/lys/index.html).
- I also added statements regarding the target audience in the manuscript (line 50).
