### 2023-10-06 lys-devel/lys#12 @pr4deepr
It launched without any Qt errors!
Closing the issue for now!
