### 2023-10-06 lys-devel/lys#13 @pr4deepr
paper comments target audience and napari
Hi @lys-devel

Great software..

In the paper, the **target audience hasn't been stated clearly**.
I know you mention its low code, but it really depends on your target audience.
Many disciplines have multidimensional data, so it would be great to give a general idea of the target audience.
This may be a low code solution to physicists with coding experience, but to a person with a background in biology and intermediate coding experience, this is more than just 'low code'.

I think you do need to have some knowledge of the respective libraries and use cases.

Have you seen [napari](https://napari.org/stable/)? It is worth mentioning it in your paper and how different lys is. You can run IPython console in it and visualize it at the same time. Its mostly aimed for visualization but there is an ecosystem of [plugins](https://www.napari-hub.org/).  It can also handle large dask arrays and xarrays).

Cheers
Pradeep
