### 2023-10-06 lys-devel/lys#14 @pr4deepr
Loading multidimensional image in lys
Hi @lys-devel
   When you say multidimensional data, I'm used to working with 3D or 4D microscopy image datasets.
For example, this data from scikit-image: https://scikit-image.org/docs/stable/api/skimage.data.html#skimage.data.cells3d

Could you add an example of how to load such an image and perhaps perform smoothing?

This is how I started in the proc.py:

```
from lys import *

from skimage import io, data
import numpy as np

def load_img():
    img = data.cells3d() #data is (60, 2, 256, 256) (Z,C,Y,X)
    print(img.shape)
    img = np.moveaxis(img, 0, 3)
    img = np.moveaxis(img, 0, 3) # make axes to (256,256,2,60)
    print(img.shape)
    return img

```

ALso, with visualization, how do I setup multicut to toggle or visualize both channels?
i.e., in the canvas I want to visualize XY, have a line graph or slider at bottom so I can scroll through Z.
Be able to look at C dimension index 0 and index 1.

Let me know if this isn't clear.

Cheers
Pradeep

https://github.com/openjournals/joss-reviews/issues/5869
