### 2023-10-06 lys-devel/lys#11 @pr4deepr
Thanks!
Could you add operating system info as well?
