### 2023-10-06 lys-devel/lys#11 @pr4deepr
Its a minor request.. otherwise i'm happy with this.
