### 2023-10-06 #5869 @pr4deepr
Hi @emdupre @kuadrat

I made 2 issues here:

https://github.com/lys-devel/lys/issues/14

https://github.com/lys-devel/lys/issues/13

Just wanted to make sure these are still  in the scope of the review?

Cheers
Pradeep
