### 2023-10-10 b03142c
Add cells3d example #14
