### 2023-10-10 lys-devel/lys#14 @lys-devel
Thank you for the comment, @pr4deepr .

I added the [example page](https://lys-devel.github.io/lys/tutorials/cells3d.html) in the document that shows how to visualize/smooth cells3d data.
I recommend you to check [basic tutorial](https://lys-devel.github.io/lys/tutorials/tutorial.html#basic-required-for-all-users) before reading the example above, if you have not read it.

Please let me know if you have any questions and comments regarding the cells3d example.
