### 2023-10-11 #5869 @emdupre
> Just wanted to make sure these are still in the scope of the review?

Thanks for checking in on this, @pr4deepr ! I see https://github.com/lys-devel/lys/issues/13 and https://github.com/lys-devel/lys/issues/14 as in-scope for the review, targeting `A statement of need:` / `State of the field: ` and ` Example usage:`, respectively. If I've misunderstood your concern, though, please let me know !

And of course, if you have other questions, please don't hesitate to ask !
