+++ paper.bib
@@ -232,3 +232,15 @@ author = {Kevin Kramer and Johan Chang},
 title = {Visualization of Multi-Dimensional Data -- The data-slicer Package}, 
 journal = {Journal of Open Source Software} 
 }
+
+@article{chiu_clack_2022, 
+title={napari: a Python Multi-Dimensional Image Viewer Platform for the Research Community}, 
+volume={28}, 
+DOI={10.1017/S1431927622006328}, 
+number={S1}, 
+journal={Microscopy and Microanalysis}, 
+publisher={Cambridge University Press}, 
+author={Chiu, Chi-Li and Clack, Nathan}, 
+year={2022}, 
+pages={1576–1577}
+}
