### 2023-10-13 lys-devel/lys#13 @lys-devel
Thank you for the fruitful comment.

I have tried `napari` as you suggested. After understanding what other similar software can do, I noticed that `lys` is a general platform for data analysis and visualization (not a simple image viewer/analysis software).
Therefore I added the paragraph that describes above point and spcify target audience as in line 51-56 in the [new manuscript](https://github.com/openjournals/joss-papers/blob/joss.05869/joss.05869/10.21105.joss.05869.pdf). In addition, I also compare `lys` with other similar software including `napari` in line 101-113.

I hope these modifications meet the publication criteria in JOSS.
