### 2023-10-20 #5869 @emdupre
👋 HI everyone, and happy Friday !

@ziatdinovmax I just wanted to follow-up and confirm if you're having any difficulties in working through your reviewer checklist, as I noticed you haven't yet been able to assess the "functionality", "documentation" or "software paper" categories -- which I know are more involved !

If you have any concerns, please let me know. And thank you again for agreeing to review _lys_ !
