### 2023-10-20 #5869 @ziatdinovmax
@lys-devel A minor suggestion re: automated tests. It is typically helpful to display the test coverage using e.g. [codecov](https://about.codecov.io/) which is free. It also signals the maturity of the package and of maintenance effort which helps attract more users.
