### 2023-10-20 #5869 @ziatdinovmax
For the statement of need, I feel it needs a further justification of why research scientists can't just rely on the already exisiting domain-specific software tools for data analysis and visualization.
