### 2023-10-21 #5869 @ziatdinovmax
@lys-devel - my apologies if I missed this in the documentation, but I think currently it doesn't account for a pretty common situation when multi-dimensional data is in a proprietary format, and one needs to use some form of a translator to convert it to lys-readable format. This seems like the first a most important step without which one cannot use this package. I suggest mentioning it and perhaps providing links to some common translator packages.
