### 2023-10-26 lys-devel/lys#15 @lys-devel
How to check coverage
A suggestion regarding the code coverage is raised in openjournals/joss-reviews#5869.

> A minor suggestion re: automated tests. It is typically helpful to display the test coverage using e.g. [codecov](https://about.codecov.io/) which is free. It also signals the maturity of the package and of maintenance effort which helps attract more users.
