### 2023-10-26 lys-devel/lys#16 @lys-devel
Statement of need of the JOSS manuscript
The modification of statement of need is raised in openjournals/joss-reviews#5869.

> For the statement of need, I feel it needs a further justification of why research scientists can't just rely on the already exisiting domain-specific software tools for data analysis and visualization.
