### 2023-10-26 lys-devel/lys#17 @lys-devel
How to load proprietary format data
The suggesting regarding the loading proprietary format data is raised in openjournals/joss-reviews#5869.

>I think currently it doesn't account for a pretty common situation when multi-dimensional data is in a proprietary format, and one needs to use some form of a translator to convert it to lys-readable format. This seems like the first and most important step without which one cannot use this package. I suggest mentioning it and perhaps providing links to some common translator packages.
