### 2023-10-26 lys-devel/lys#15 @lys-devel
Hi @ziatdinovmax,

I added the statement regarding the code coverage in the [document](https://lys-devel.github.io/lys/tutorials/test.html).
Thank you for your kind suggestion.
