### 2023-10-27 lys-devel/lys#17 @pr4deepr
@ziatdinovmax, do you think this can be addressed by having an example in the 'examples gallery'. similar to [this](https://lys-devel.github.io/lys/tutorials/cells3d.html) ?
Also, would you have any suggestions on a proprietary data format to use?
