### 2023-10-27 a64df6b
Add loading data examples in documents #17
