### 2023-10-27 lys-devel/lys#17 @lys-devel
@ziatdinovmax, I made examples to translate several format data in [documents](https://lys-devel.github.io/lys/tutorials/load_examples.html).
If you think there is another important format to be adressed, please let me know.
