### 2023-10-27 lys-devel/lys#16 @lys-devel
@ziatdinovmax, I added the sentences to justify the above points in line 22-32 in the [new manucript](https://github.com/openjournals/joss-papers/blob/joss.05869/joss.05869/10.21105.joss.05869.pdf).
I thank you for the comments, and I hope these modifications meet the publication criteria in JOSS.
