+++ paper.md
@@ -24,11 +24,11 @@ bibliography: paper.bib
 
 # Statement of need
 
-Data analysis and visualization are indispensable parts of scientific research. Understanding the experimental and simulated data deeply is essential for extracting complex phenomena behind them. To this end, intuitive and fast analysis/visualization is highly required. There are  several well-known softwares that focus on analysis and visualization, such as Igor Pro (WaveMetrics) and Matlab (MathWorks) as well as the domain-specific softwares in respective fiels. Although they have been utilized for scientific research for a long time, recent progress in experimental/computational methods change the situation. The data sizes obtained from the experiments/calculations are acceleratedly increased within this decades, reaching more than terabytes in many fields. In addition, sophisticated analysis methods for such a huge data has also been matured. As a result, general analysis ecosystem maintained by the large scientific open source community has an advantage in the state-of-the-art analysis as compared to the proprietary/domain-specific softwares. Python and related scientific libraries are one of the most popular analysis/visualization environment at the moment. In the scientific Python ecosystem, flexible and fast analysis of numerical arrays has been done by several popular libraries such as `numpy`[@harris2020array], `scipy`[@2020SciPy-NMeth], `dask`[@rocklin2015dask] combined with visualization tools such as `matplotlib`[@Hunter:2007], `pyqtgraph`[@pyqtgraph] and `Mayavi`[@ramachandran2011mayavi]. Utilizing these variety of libraries, sophisticated data analysis/visualization tools such as `data-slicer` [@Kramer2021] and `napari`[@chiu_clack_2022] have been developed. Development of `Jupyter Notebook` [@Kluyver2016jupyter] and related libraries further enhance the capability of interactive data analysis. However, most of these libraries require users to be familiar with low-level application programming interfaces (APIs), which prevents intuitive analysis and visualization. In particular, when the analyzed data is a more than 3-dimensional array, even simple interactive visualization usually requires Python code of tens of lines. Such on-demand analysis and visualization programs should be modified (and tested) when the data change, although a very similar process is frequently applied to data with different dimensions. Furthermore, the code for such an analysis should be preserved to guarantee the reproducibility of scientific results.
[…]
