### 2023-11-06 lys-devel/lys#20 @lys-devel
Manuscript modification
Several concern regarding the manuscript is raised in openjournals/joss-reviews#5869.

Examples of typo and grammar or sentence structure:

- Line 25: field
- Line 26 to 31: grammar and sentences can be condensed and made clearer.
- Line 52 to 54: unclear
- Line 57: Grammar. An example (not the best one): The philosophy of lys is to serve as a versatile data analysis and visualization platform, rather than a basic image viewer/analysis program. It is tailored to research scientists who would like to analyze and visualize their data with minimal coding.
- Line 62: significantly reduces the time
- Line 105: expansion capability -> you could use the term modularity maybe?
- Line 110: Users can verify the software and modify if needed

Other:
- Line 112. I would not say interactive GUI is lacking in the scientific ecosystem. napari can be used to visualize images and it also has an ipython console where you can enter code to process data interactively.
- Line 117. napari can be extended for data analysis. I think you should clarify and say lys is 'general purpose', i.e. it can handle different data types perhaps and not restricted to image data only. For example, napari has a massive ecosystem of plugins: https://www.napari-hub.org/
