### 2023-11-06 lys-devel/lys#20 @lys-devel
@pr4deepr, thank you for the comments.
I will proofread the manuscript with my colleague. Before that, I have modified the manuscript following your comments ([new manuscript](https://github.com/openjournals/joss-papers/blob/joss.05869/joss.05869/10.21105.joss.05869.pdf)) as following.

**Grammar/Typo**
- Line 25 in the new manuscript: I fixed a typo (field)
- Line 26 to 30: I modified the sentences below.
> The data sizes obtained from experiments and calculations have increased rapidly over the past decade, reaching more than
 terabytes in many fields. As a result, the general analysis ecosystem maintained by the large scientific open-source community has an advantage in state-of-the-art analysis compared to proprietary/domain-specific software.
- Line 51 to 53: I modified the sentences below.
>  In particular, a tool for interactive and fast analysis of multi-dimensional arrays has been developed. This tool allows all analysis processes to be exported as a single file, ensuring scientific reproducibility
- Line 57-59: I modified the sentences below.
> The philosophy of lys is to serve as a versatile data analysis and visualization platform, rather than a basic image viewer/analysis program. It is developed with the aim of minimizing the time required for data analysis and visualization for researchers.
- Line 62: I replaced "compress" with "reduce".
- Line 106: I replaced "expansion capability" with "extensibility" since the word "modularity" was slightly different from what I meant.
- Line 111: I modified the sentences below.
> Users can verify the software and mofify if needed, which cannot be realized in proprietary softwares.

**Other**
-  _I would not say interactive GUI is lacking in the scientific ecosystem. napari can be used to visualize images and it also has an ipython console where you can enter code to process data interactively._

I agree that it was incorrect statement, I modified the sentences below to correctly state the situation in Line 113.
> Although there have been much effort to realize sophisticated GUIs such as napari(Chiu & Clack, 2022) and data-slicer(Kramer & Chang, 2021), it is still very limited in the scientific Python ecosystems so far.

- _napari can be extended for data analysis. I think you should clarify and say lys is 'general purpose', i.e. it can handle different data types perhaps and not restricted to image data only. For example, napari has a massive ecosystem of plugins: https://www.napari-hub.org/_

Thank you for the suggestion. I modified the sentences below to compare `lys` with `napari` in Line 121.
>Although napari also offers similar functionarities for multi-dimensional images, lys is designed to handle variety of data types (images, curves, contours, vector fields, and RGB images) to serve as general-purpose platform.
