### 2023-11-10 lys-devel/lys#20 @lys-devel
@pr4deepr, I have proofread the manuscript with my colleagues [new manuscript](https://github.com/openjournals/joss-papers/blob/joss.05869/joss.05869/10.21105.joss.05869.pdf). Thanks to your comments, I think the manuscript is much improved. Thank you very much.
