### 2023-11-14 #5869 @emdupre
Hi everyone,

I'm back in office ! Thank you for your patience.

@pr4deepr and @ziatdinovmax, thank you for your confirmation of your reviews !
@pr4deepr, I noted that https://github.com/lys-devel/lys/issues/20 was closed following your comments, so I believe these are now addressed. Please let me know, though, if this is not the case.

Otherwise, it looks like all three reviewer checklists are now completed ! @kuadrat could you please confirm that you are happy with this submission for `lys` in its current form ? I noted that you still have two open issues, but it looks like you've marked these as "not review-critical," so I wanted to confirm.

Once I have final confirmation, I'll perform a few additional editorial checks to help in processing this submission  🚀
