### 2023-11-14 #5869 @kuadrat
Welcome back, @emdupre!

Indeed, I also think that the manuscript and the software fulfill the conditions for publication in JOSS. The remaining open issues are not review critical.
