### 2023-11-21 #5869 @emdupre
Thank you for your patience, @lys-devel ! I've reviewed and am also happy with the software 🎉 I have a few editorial requests on the software paper itself.

Major comments
- The summary currently points to several specific software packages and kinds of computing environments. While this is very useful for regular developers, it may be less familiar to non-specialist audiences. The [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) for the Summary ask that this section focus on:
  > describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

  Could you thus please re-orient the Summary towards this non-specialist audience ? I would suggest to use similar language to e.g., L20-22 and L58-64.
- Figure 1. Can you please describe each of the figure panels directly in the figure caption, rather than in-text ?
- Figure 2. Similarly, can you please expand the figure caption ? I understand that it is not entirely possible to move the description into the caption given its length.
- napari and data-slicer are first mentioned on L37 as "sophisticated data analysis/visualization tools" ; however, a more direct comparison of lys with these tools is not provided until L115. It would be useful to have this discussion moved into the Statement of Need, to make the current state of the field clearer to the reader. Per the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), the Statement of Need:
  > clearly illustrates the research purpose of the software and places it in the context of related work.

  If desired, discussion of L47 - L57 could be moved into the Overview to focus each section's content.

Minor comments
- L9 : Please note that "GUI interface" is redundant. You can simply say 'GUI' or 'interface'
- L21 : Please remove 'the' before 'experimental and simulated data'
- L23 : Please capitalize Igor Pro as IGOR Pro
- L24 : Please capitalize Matlab as MATLAB
- L24 : Please add RRID for [IGOR Pro](https://scicrunch.org/resolver/SCR_000325) and [MATLAB](https://scicrunch.org/resolver/SCR_001622).
- L33 : Please add spaces before each citation and its corresponding library
- L75 : preprocess --> preprocessing
- Figure 1. Could you please explicitly note that this GUI is the `MultiCut` interface referred to throughout the text ?
