+++ paper.md
@@ -30,8 +30,6 @@ The multi-dimensional data analysis and visualization platform, `lys`, offers a 
 
 The philosophy of lys is to serve as a versatile data analysis and visualization platform, rather than a basic image viewer/analysis program. It was developed to minimize the time required for data analysis and visualization for researchers. All of the processes from loading data to generating publication-quality figures can be done in `lys`. In addition, these processes (including user-defined Python code) are stored in a single directory and can be used to reproduce the results. The rich features of `lys` significantly reduce the time for analysis/visualization of multi-dimensional arrays.
 
-![Screenshot of lys. Users can execute arbitrary Python commands from embedded CUI (#1), which can be extended by the user-defined scripts (#2). Matplotlib figures that contain curves, images, vector fields, and RGB images can be displayed (#3). These figures can be edited via GUI in the sidebar (#4).](Fig1.png)
-
 # Overview
 
 `lys` is a hybrid GUI/CUI platform oriented towards multi-dimensional data analysis. Figure 1 shows the main features of `lys`. Arbitrary Python commands can be executed from an integrated Python shell (#1). User-defined Python scripts can be edited by the internal editor (#2) and can be executed. Matplotlib graphs that contain curves, images, vector fields, and RGB images can be displayed (#3) and edited via GUI in the sidebar (#4). 
@@ -40,12 +38,14 @@ The philosophy of lys is to serve as a versatile data analysis and visualization
 
 $$A'(x_i,y_j,t_k) = \mathcal{M}[A(x_i,y_j,t_k)],$$
 
+![Screenshot of lys. Users can execute arbitrary Python commands from embedded CUI (#1), which can be extended by the user-defined scripts (#2). Matplotlib figures that contain curves, images, vector fields, and RGB images can be displayed (#3). These figures can be edited via GUI in the sidebar (#4).](Fig1.png)
+
 where $P = \mathcal{M}$ represents median filter to remove the noise. Since the median filter does not change the dimension of data, $N=3$ equals $M$ in this case. The preprocessing is used for heavy analysis that requires whole $N$-dimensional data. Second, MultiCut generates an arbitrary number of 2-dimensional images and 1-dimensional curves from $M$-dimensional $A'$ by taking a summation along arbitrary axes. In the example case in Fig. 2, an image $I(x_i,y_j)$ and a curve $C(t_k)$ is generated as:
 
 $$
 \begin{aligned}
 I(x_i,y_j) &= \sum_{t_k}A'(x_i,y_j,t_k), \\
-C(t_k) &= \sum_{x_i, y_i}A'(x_i,y_j,t_k).
+C(t_k) &= \sum_{x_i, y_j}A'(x_i,y_j,t_k).
 \end{aligned}
 $$
 
@@ -61,7 +61,7 @@ Different from the preprocessing (step 1), the postprocessing function accesses 
 
 In addition to the features described above, `lys` provides some basic analysis such as data fitting and array editor GUIs. Combining these functionalities of `lys` offers intuitive, low-code, fast, and flexible analysis to users not familiar with Python while preserving the extensibility for experts.
 
[…]
