### 2023-11-24 #5869 @lys-devel
Thank you for you editorial effort, @emdupre. I also appreciate all reviewers for their kind comments. I modified the manuscript following your advice.
- Summary
> Could you thus please re-orient the Summary towards this non-specialist audience ?

Thank you for the comment. I modified the Summary towards non-specialist audience in Line 6-20.

- Figure captions
> Figure 1. Can you please describe each of the figure panels directly in the figure caption, rather than in-text ?
> Figure 2. Similarly, can you please expand the figure caption ? I understand that it is not entirely possible to move the description into the caption given its length.

I modified the figure captions of Fig. 1 and 2.

- Statement of need
> napari and data-slicer are first mentioned on L37 as "sophisticated data analysis/visualization tools" ; however, a more direct comparison of lys with these tools is not provided until L115. It would be useful to have this discussion moved into the Statement of Need, to make the current state of the field clearer to the reader

Thank you for the important comment. I added the sentences that describe the current state of the field in Line 46-54.
I did not change the discussion in Line 117-133 because I consider the comparison between other software is also needed after describing what is lys.

- Other
> L24 : Please add RRID for [IGOR Pro](https://scicrunch.org/resolver/SCR_000325) and [MATLAB](https://scicrunch.org/resolver/SCR_001622).

I could not find example of how to add RRID in JOSSS website, so I added RRID directly in Line 26. Could you confirm this style of RRID is suitable for JOSS?

> Figure 1. Could you please explicitly note that this GUI is the MultiCut interface referred to throughout the text ?

Figure 1 is not the screenshot of MultiCut (there is no screenshot of MultiCut in the manuscript). So I did not modify the manuscript regarding this point.

I hope the manuscript has been modified suitably for JOSS, Thank you.
