### 2023-11-29 #5869 @emdupre
Thank you, @lys-devel ! I appreciate you making these changes.

> I could not find example of how to add RRID in JOSSS website, so I added RRID directly in Line 26. Could you confirm this style of RRID is suitable for JOSS?

Apologies for not providing this ; if you could please remove the space between `RRID:` and the SCR identifier, these would be in the correct format. Thus, MATLAB would be `RRID:SCR_001622`.

The only remaining concern I see is that the software are stylized as `Igor Pro` and `Matlab` on L120. Could you please correct the capitalization there as well ?
