### 2023-12-01 b40c8b5
Fix manuscript
