### 2023-12-01 #5869 @lys-devel
Thank you, @emdupre. I have fixed the manuscript and register the new release in zenodo.
Tagged release: [v0.3.3](https://github.com/lys-devel/lys/releases/tag/v0.3.3)
DOI of the archive: [10.5281/zenodo.10241638](https://doi.org/10.5281/zenodo.10241638)

Since I made release v0.3.3 before enabling Zenodo github integration, I made duplicated release [v0.3.3_zenodo](https://github.com/lys-devel/lys/releases/tag/v0.3.3_zenodo), which is registered in Zenodo (v0.3.3 is NOT registered in Zenodo). Please use this release if it is preffered.
