### 2023-12-01 #5869 @emdupre
Thank you, @lys-devel ! I'm confirming a few additional points in the post-review checklist, above 👆

One small issue with the archive : It looks like its license is currently listed as CC-BY 4.0, rather than the GPL 3.0 license of _lys_. Could you please correct this ? You should be able to do so by [editing the existing archive's metadata](https://help.zenodo.org/faq/#technical).

If it requires minting a new DOI, please let me know !
