+++ paper.md
@@ -20,19 +20,19 @@ bibliography: paper.bib
 
 # Summary
 
-Analyzing and visualizing scientific data is an essential part of scientific research, by which researchers investigate complex phenomena behind it. In particular, the recent development of an open-source scientific Python ecosystem enables us to utilize state-of-the-art analysis and visualization. However, investigation of multi-dimensional data array still requires substantial coding, preventing intuitive and flexible analysis/visualization. `lys` is a Python-based multi-dimensional data analysis and visualization platform, which provides graphical user interfaces (GUI) to intuitively and flexibly manipulate multi-dimensional data arrays and publication-quality graphics. Massive multi-dimensional data over hundreds of gigabytes can be analyzed via automatic parallel calculation behind the GUI when `lys` is run in high-performance computers (HPCs). As well as the user-friendly GUIs, `lys` also provides flexibility for experts through its character user interface (CUI). The hybrid GUI/CUI architecture in `lys` enables an intuitive, low-code, parallel, flexible, and extensible multi-dimensional data analysis. `lys` is designed as a versatile data analysis and visualization platform that can handle all analysis processes from data loading to publication-quality figure generation. These features of `lys` enable us to minimize the time required for data analysis and visualization for a broad range of users.
+Analyzing and visualizing scientific data is an essential part of scientific research, by which researchers investigate complex phenomena behind it. In particular, the recent development of an open-source scientific Python ecosystem enables us to utilize state-of-the-art analysis and visualization. However, investigation of multi-dimensional data array still requires substantial coding, preventing intuitive and flexible analysis/visualization. `lys` is a Python-based multi-dimensional data analysis and visualization platform, which provides graphical user interfaces (GUIs) to intuitively and flexibly manipulate multi-dimensional data arrays and publication-quality graphics. Massive multi-dimensional data over hundreds of gigabytes can be analyzed via automatic parallel calculation behind the GUI when `lys` is run in high-performance computers (HPCs). As well as the user-friendly GUIs, `lys` also provides flexibility for experts through its character user interface (CUI). The hybrid GUI/CUI architecture in `lys` enables an intuitive, low-code, parallel, flexible, and extensible multi-dimensional data analysis. `lys` is designed as a versatile data analysis and visualization platform that can handle all analysis processes from data loading to publication-quality figure generation. These features of `lys` enable us to minimize the time required for data analysis and visualization for a broad range of users.
 
 # Statement of need
 
[…]
