### 2023-12-13 #5869 @emdupre
Thank you for your engagement throughout the review process, @lys-devel ! I'm now happy to recommend _lys_ to the EiC team for publication in JOSS 🎉

Thank you, too, to @pr4deepr, @kuadrat, and @ziatdinovmax for your reviews and input throughout this process ! JOSS runs thanks to volunteers like you, so thank you 💐
