### 2023-12-13 #5869 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
