+++ paper.bib
@@ -95,7 +95,7 @@ Year = {2016}
 
 @misc{pyqtgraph,
   author = {L. Campagnola, K. Lyons, and O. Moore},
-  title = {pyqtgraph: Scientific Graphics and GUI Library for {P}ython},
+  title = {{PyQtGraph}: Scientific Graphics and GUI Library for {P}ython},
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -164,7 +164,7 @@ eprint = { https://doi.org/10.1021/acs.nanolett.2c03938}}
 
 @article{SciAdv2021,
 author = {Takahiro Shimojima  and Asuka Nakamura  and Xiuzhen Yu  and Kosuke Karube  and Yasujiro Taguchi  and Yoshinori Tokura  and Kyoko Ishizaka },
-title = {Nano-to-micro spatiotemporal imaging of magnetic skyrmion\&\#x2019;s life cycle},
+title = {Nano-to-micro spatiotemporal imaging of magnetic skyrmion's life cycle},
 journal = {Science Advances},
 volume = {7},
 number = {25},
