### 2023-12-13 #5869 @danielskatz
@lys-devel - I'm the track editor, working on the final steps for accepting and publishing your submission. I've suggested some changes in https://github.com/lys-devel/lys/pull/22. Please merge this, or let me know what you disagree with. Also, I note your use of RRIDs, in the statement of need section. JOSS doesn't typically use RRIDs, and while I don't mind you using them, it would be useful if you would also use more widely-understood citations. Can you replace or add references for these items?
