### 2023-12-13 #5869 @emdupre
Hi @danielskatz, thanks for the quick follow-up !

Just to flag that I had requested RRIDs for IGOR Pro and MATLAB as the previous citations had only listed the associated companies. So, this was on me, not @lys-devel. I traditionally find RRIDs clearer in this case, but will defer to you as the expert on these citations !
