### 2023-12-13 #5869 @danielskatz
Let's use both then, as I don't think many people outside of the life sciences have ever heard of RRIDs, let alone know what to do with them.  (https://f1000research.com/articles/9-1257/v2 might be useful in terms of guidance about how to cite software that's not clear otherwise)
