### 2023-12-13 #5869 @emdupre
Thanks, @danielskatz !

@lys-devel , just to make the updates a bit easier, here is the relevant MATLAB citation in bibtex :
```
@software{MATLAB,
year = {2023},
author = {The MathWorks Inc.},
title = {MATLAB (R2023b)},
publisher = {The MathWorks Inc.},
address = {Natick, Massachusetts, United States},
url = {https://www.mathworks.com}
}
```
I've suggested citing the most recent version, but please amend if necessary.

For IGOR Pro, it seems [the field-accepted standard](https://www.wavemetrics.com/forum/general/citing-igor-pro-paper) is `(WaveMetrics, Inc., Lake Oswego, OR, USA)`.
