+++ paper.md
@@ -24,7 +24,7 @@ Analyzing and visualizing scientific data is an essential part of scientific res
 
 # Statement of need
 
-Data analysis and visualization are indispensable parts of scientific research. Understanding experimental and simulated data deeply is essential for extracting complex phenomena behind them. To this end, intuitive and fast analysis/visualization is highly required. There are several well-known software that focus on analysis and visualization, such as `IGOR Pro` (WaveMetrics, RRID:SCR_000325) and `MATLAB` (MathWorks, RRID:SCR_001622), as well as domain-specific software in respective fields. Although they have been utilized for scientific research for a long time, recent progress in experimental/computational methods has changed the situation. The data sizes obtained from experiments and calculations have increased rapidly over the past decade, reaching more than terabytes in many fields. As a result, the general analysis ecosystem maintained by the large scientific open-source community has an advantage in state-of-the-art analysis compared to proprietary/domain-specific software. Python and its scientific libraries provide one of the most popular analysis/visualization environments at the moment. In the scientific Python ecosystem, flexible and fast analysis of numerical arrays has been done by several popular libraries such as `numpy` [@harris2020array], `scipy` [@2020SciPy-NMeth], `dask` [@rocklin2015dask] combined with visualization tools such as `matplotlib` [@Hunter:2007], `pyqtgraph` [@pyqtgraph] and `Mayavi` [@ramachandran2011mayavi]. Development of `Jupyter Notebook` [@Kluyver2016jupyter] and related libraries further enhance the capability of interactive data analysis. However, most of these libraries require users to be familiar with low-level application programming interfaces (APIs), which prevents intuitive analysis and visualization. In particular, when the analyzed data is a more than 3-dimensional array, even simple interactive visualization usually requires Python code of tens of lines. Such on-demand analysis and visualization programs should be modified (and tested) when the data change, although a very similar process is frequently applied to data with different dimensions. Furthermore, the code for such an analysis should be preserved to guarantee the reproducibility of scientific results. To solve these problems, several Python-based tools such as  `data-slicer` [@Kramer2021] and `napari` [@chiu_clack_2022] have been developed. Although both of them offer sophisticated GUIs, they are not designed as a general research platform. `data-slicer` focuses on quick data inspection and visualization rather than publication-quality figure generation. `napari` is mainly developed for multi-dimensional image data, and therefore visualization of other data types such as curves, contours, vectors is limited. In such situations, a versatile Python-based analysis and visualization platform that can handle all scientific analysis processes, from data loading to publication-quality figure generation of a variety of data types, is highly required.
[…]
