+++ paper.bib
@@ -244,3 +244,21 @@ author={Chiu, Chi-Li and Clack, Nathan},
 year={2022}, 
 pages={1576–1577}
 }
+
+@software{MATLAB,
+year = {2023},
+author = {The MathWorks Inc.},
+title = {MATLAB (R2023b)},
+publisher = {The MathWorks Inc.},
+address = {Natick, Massachusetts, United States},
+url = {https://www.mathworks.com}
+}
+
+@software{IGOR,
+year = {2023},
+author = {WaveMetrics Inc.},
+title = {IGOR Pro},
+publisher = {WaveMetrics Inc.},
+adress = {Lake Oswego, Oregon, United States},
+url = {https://www.wavemetrics.com}
+}
