+++ paper.bib
@@ -247,7 +247,7 @@ pages={1576–1577}
 
 @software{MATLAB,
 year = {2023},
-author = {The MathWorks Inc.},
+author = {{The MathWorks Inc.}},
 title = {MATLAB (R2023b)},
 publisher = {The MathWorks Inc.},
 address = {Natick, Massachusetts, United States},
@@ -256,7 +256,7 @@ url = {https://www.mathworks.com}
 
 @software{IGOR,
 year = {2023},
-author = {WaveMetrics Inc.},
+author = {{WaveMetrics Inc.}},
 title = {IGOR Pro},
 publisher = {WaveMetrics Inc.},
 adress = {Lake Oswego, Oregon, United States},
