+++ paper.bib
@@ -257,7 +257,7 @@ url = {https://www.mathworks.com}
 @software{IGOR,
 year = {2023},
 author = {{WaveMetrics Inc.}},
-title = {IGOR Pro},
+title = {IGOR Pro 9.05},
 publisher = {WaveMetrics Inc.},
 adress = {Lake Oswego, Oregon, United States},
 url = {https://www.wavemetrics.com}
