### 2023-12-14 #5869 @lys-devel
Hi @emdupre, @danielskatz!

Thank you for your kind suggestions. I have merged the [PR #22](https://github.com/lys-devel/lys/pull/22) and have added citation for MATLAB and IGOR Pro.
Could you confirm these are suitable for JOSS?
