### 2023-12-14 #5869 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
